# genai/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Financial News Gen AI Service
  description: Enrichment API for financial news articles.
  version: 0.1.0
paths:
...
  /api/v1/news/intake:
    post:
      tags:
      - ingestion
      summary: Store raw news metadata and queue enrichment
      operationId: ingest_raw_news_api_v1_news_intake_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RawNewsIngestionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionAcceptedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/news/intake-text:
    post:
      tags:
      - ingestion
      summary: Store licensed article or summary text and queue enrichment
      operationId: ingest_raw_news_text_api_v1_news_intake_text_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectTextIngestionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionAcceptedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/news/{news_id}:
    get:
      tags:
      - ingestion
      summary: Get raw news, latest job, and enrichment result
      operationId: get_news_status_api_v1_news__news_id__get
      parameters:
      - name: news_id
        in: path
        required: true
        schema:
          type: string
          title: News Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsProcessingStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/news/{news_id}/result:
    get:
      tags:
      - ingestion
      summary: Get backend-friendly enrichment result for a news item
      operationId: get_news_result_api_v1_news__news_id__result_get
      parameters:
      - name: news_id
        in: path
        required: true
        schema:
          type: string
          title: News Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsResultResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AnalysisOutcome:
      type: string
      enum:
      - success
      - partial_success
      - fatal_failure
      title: AnalysisOutcome
    AnalysisStatus:
      type: string
      enum:
      - pending
      - fetch_failed
      - clean_failed
      - validate_failed
      - summarize_failed
      - sentiment_failed
      - xai_failed
      - mixed_detection_failed
      - build_payload_failed
      - persist_failed
      - completed_with_partial_results
      - completed
      title: AnalysisStatus
    ArticleEnrichmentRequest:
      properties:
        news_id:
          type: string
          minLength: 1
          title: News Id
          description: Unique news identifier.
        title:
          type: string
          minLength: 1
          title: Title
          description: Original article title.
        link:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Link
          description: Canonical article URL.
        ticker:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Ticker
          description: Optional list of related ticker symbols.
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
          description: Publisher or source name.
        published_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published At
          description: Original publication timestamp.
      additionalProperties: false
      type: object
      required:
      - news_id
      - title
      - link
      title: ArticleEnrichmentRequest
    ArticleEnrichmentResponse:
      properties:
        news_id:
          type: string
          title: News Id
          description: Unique news identifier.
        title:
          type: string
          title: Title
          description: Original article title.
        link:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Link
          description: Canonical article URL.
        summary_3lines:
          items:
            $ref: '#/components/schemas/SummaryLine'
          type: array
          maxItems: 3
          title: Summary 3Lines
          description: Up to three grounded summary lines when available.
        sentiment:
          anyOf:
          - $ref: '#/components/schemas/SentimentResult'
          - type: 'null'
          description: Article sentiment output when analysis succeeded.
        xai:
          anyOf:
          - $ref: '#/components/schemas/XAIPayload'
          - type: 'null'
          description: Explainability payload when analysis succeeded.
        mixed_flags:
          anyOf:
          - $ref: '#/components/schemas/MixedConflictPayload'
          - type: 'null'
          description: Mixed/conflict analysis output.
        status:
          $ref: '#/components/schemas/EnrichmentStatus'
          description: Overall pipeline status.
        outcome:
          type: string
          enum:
          - success
          - partial_success
          - fatal_failure
          title: Outcome
          description: Top-level outcome separating success, partial success, and
            fatal failure.
        analyzed_at:
          type: string
          format: date-time
          title: Analyzed At
          description: When analysis was completed or last updated.
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
          description: Top-level pipeline error, if any.
        stage_statuses:
          items:
            $ref: '#/components/schemas/InternalStageStatus'
          type: array
          title: Stage Statuses
          description: Optional internal execution status per pipeline stage.
      additionalProperties: false
      type: object
      required:
      - news_id
      - title
      - link
      - status
      - outcome
      title: ArticleEnrichmentResponse
...
    ArticleFetchResult:
      properties:
        link:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Link
          description: Requested article URL.
        publisher_domain:
          type: string
          title: Publisher Domain
          description: Publisher domain derived from the requested article URL.
          default: ''
        final_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Final Url
          description: Final resolved URL after redirects, when available.
        http_status_code:
          anyOf:
          - type: integer
            maximum: 599.0
            minimum: 100.0
          - type: 'null'
          title: Http Status Code
          description: HTTP status code associated with the fetch failure, when available.
        content_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Type
          description: Response content type observed during fetch, when available.
        extraction_source:
          anyOf:
          - $ref: '#/components/schemas/ArticleTextSource'
          - type: 'null'
          description: Which parser path produced the extracted article text, when
            available.
        attempt_count:
          type: integer
          minimum: 1.0
          title: Attempt Count
          description: Number of fetch attempts made before returning the result.
          default: 1
        raw_text:
          type: string
          title: Raw Text
          description: Text extracted directly from the fetched article HTML.
          default: ''
        cleaned_text:
          type: string
          title: Cleaned Text
          description: Parser-cleaned article text for downstream Gen AI processing.
          default: ''
        fetch_status:
          $ref: '#/components/schemas/ArticleFetchStatus'
          description: Final status of the article fetch and parse flow.
        retryable:
          type: boolean
          title: Retryable
          description: Whether the fetch failure appears transient and worth retrying
            later.
          default: false
        failure_category:
          anyOf:
          - $ref: '#/components/schemas/ArticleFetchFailureCategory'
          - type: 'null'
          description: Normalized failure category for analytics and retry policy
            decisions.
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
          description: Failure reason when fetching or parsing did not succeed.
      additionalProperties: false
      type: object
      required:
      - link
      - fetch_status
      title: ArticleFetchResult
    ArticleFetchStatus:
      type: string
      enum:
      - success
      - fetch_failed
      - parse_failed
      title: ArticleFetchStatus
...
    ArticleMixedDetectionResult:
      properties:
        status:
          $ref: '#/components/schemas/MixedStatus'
          description: Overall article-level mixed status.
        is_mixed:
          type: boolean
          title: Is Mixed
          description: Whether the article is mixed.
        has_conflicting_signals:
          type: boolean
          title: Has Conflicting Signals
          description: Whether chunk-level signals conflict meaningfully.
        dominant_sentiment:
          anyOf:
          - $ref: '#/components/schemas/FinBERTSentimentLabel'
          - type: 'null'
          description: Dominant article sentiment if one remains identifiable.
        score:
          type: number
          maximum: 100.0
          minimum: -100.0
          title: Score
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
        disagreement_ratio:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Disagreement Ratio
        triggered_reason_codes:
          items:
            $ref: '#/components/schemas/ArticleMixedReasonCode'
          type: array
          title: Triggered Reason Codes
        reasons:
          items:
            $ref: '#/components/schemas/MixedReason'
          type: array
          title: Reasons
        thresholds:
          $ref: '#/components/schemas/ArticleMixedConfig'
          description: Thresholds used for evaluation.
      additionalProperties: false
      type: object
      required:
      - status
      - is_mixed
      - has_conflicting_signals
      - score
      - confidence
      - disagreement_ratio
      - thresholds
      title: ArticleMixedDetectionResult
    ArticleMixedReasonCode:
      type: string
      enum:
      - low_confidence
      - small_score_gap
      - chunk_disagreement
      title: ArticleMixedReasonCode
...
    DirectTextIngestionRequest:
      properties:
        news_id:
          type: string
          minLength: 1
          title: News Id
          description: Unique news identifier.
        title:
          type: string
          minLength: 1
          title: Title
          description: Original article title.
        link:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Link
          description: Canonical article URL.
        ticker:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Ticker
          description: Optional list of related ticker symbols.
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
          description: Publisher or source name.
        published_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published At
          description: Original publication timestamp.
        article_text:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Article Text
          description: Licensed full article text supplied directly by the upstream
            provider.
        summary_text:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Summary Text
          description: Licensed summary/snippet text supplied directly by the upstream
            provider.
      additionalProperties: false
      type: object
      required:
      - news_id
      - title
      - link
      title: DirectTextIngestionRequest
      description: Inbound raw-news payload with licensed text supplied directly by
        the upstream backend.
    EnrichmentJobRecord:
      properties:
        job_id:
          type: string
          minLength: 1
          title: Job Id
        news_id:
          type: string
          minLength: 1
          title: News Id
        status:
          $ref: '#/components/schemas/EnrichmentJobStatus'
          description: Queue processing status.
        attempts:
          type: integer
          minimum: 0.0
          title: Attempts
        max_attempts:
          type: integer
          minimum: 1.0
          title: Max Attempts
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
        last_analysis_status:
          anyOf:
          - $ref: '#/components/schemas/AnalysisStatus'
          - type: 'null'
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        next_retry_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Next Retry At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
      additionalProperties: false
      type: object
      required:
      - job_id
      - news_id
      - status
      - attempts
      - max_attempts
      - created_at
      - updated_at
      title: EnrichmentJobRecord
    EnrichmentJobStatus:
      type: string
      enum:
      - queued
      - retry_pending
      - processing
      - completed
      - failed
      title: EnrichmentJobStatus
    EnrichmentStatus:
      type: string
      enum:
      - pending
      - processing
      - completed
      - partial
      - failed
      title: EnrichmentStatus
    EnrichmentStoragePayload:
      properties:
        news_id:
          type: string
          minLength: 1
          title: News Id
        title:
          type: string
          minLength: 1
          title: Title
        link:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Link
          description: Canonical article URL.
        summary_3lines:
          items:
            type: string
          type: array
          title: Summary 3Lines
          description: Three-line article summary when available.
        sentiment:
          anyOf:
          - $ref: '#/components/schemas/StoredSentimentPayload'
          - type: 'null'
          description: Stored article-level sentiment payload.
        xai:
          anyOf:
          - $ref: '#/components/schemas/XAIResult'
          - type: 'null'
          description: Stored explainability payload.
        article_mixed:
          anyOf:
          - $ref: '#/components/schemas/ArticleMixedDetectionResult'
          - type: 'null'
          description: Article-level mixed/conflict detection result.
        ticker_mixed:
          anyOf:
          - $ref: '#/components/schemas/TickerMixedDetectionResult'
          - type: 'null'
          description: Ticker-level mixed/conflict detection result.
        analysis_status:
          $ref: '#/components/schemas/AnalysisStatus'
          description: Centralized overall pipeline status.
          default: pending
        analysis_outcome:
          $ref: '#/components/schemas/AnalysisOutcome'
          description: Top-level outcome separating success, partial success, and
            fatal failure.
          default: partial_success
        analyzed_at:
          type: string
          format: date-time
          title: Analyzed At
          description: Timestamp when this storage payload was assembled.
        cleaned_text_available:
          type: boolean
          title: Cleaned Text Available
          description: Whether cleaned article text existed at build time.
          default: false
        fetch_result:
          anyOf:
          - $ref: '#/components/schemas/ArticleFetchResult'
          - type: 'null'
          description: Optional fetch trace metadata.
        stage_statuses:
          items:
            $ref: '#/components/schemas/PipelineStageResult'
          type: array
          title: Stage Statuses
          description: Per-stage execution details for debugging and observability.
        errors:
          items:
            $ref: '#/components/schemas/StoragePayloadError'
          type: array
          title: Errors
          description: Fatal and non-fatal issues captured during payload assembly.
      additionalProperties: false
      type: object
      required:
      - news_id
      - title
      - link
      title: EnrichmentStoragePayload
...
    FinBERTSentimentLabel:
      type: string
      enum:
      - positive
      - neutral
      - negative
      title: FinBERTSentimentLabel
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IngestionAcceptedResponse:
      properties:
        news_id:
          type: string
          minLength: 1
          title: News Id
        queued:
          type: boolean
          title: Queued
          description: Whether a new job was enqueued.
        message:
          type: string
          minLength: 1
          title: Message
        job:
          $ref: '#/components/schemas/EnrichmentJobRecord'
          description: Active or newly created job record.
      additionalProperties: false
      type: object
      required:
      - news_id
      - queued
      - message
      - job
      title: IngestionAcceptedResponse
...
    MixedConflictPayload:
      properties:
        is_mixed:
          type: boolean
          title: Is Mixed
          description: Whether the article has mixed sentiment.
        has_conflicting_signals:
          type: boolean
          title: Has Conflicting Signals
          description: Whether positive and negative evidence conflict materially.
        dominant_sentiment:
          anyOf:
          - $ref: '#/components/schemas/SentimentLabel'
          - type: 'null'
          description: Dominant sentiment if one still exists.
        conflict_reasons:
          items:
            type: string
          type: array
          title: Conflict Reasons
          description: Short explanations for mixed or conflicting signals.
      additionalProperties: false
      type: object
      required:
      - is_mixed
      - has_conflicting_signals
      title: MixedConflictPayload
...
    MixedStatus:
      type: string
      enum:
      - clear
      - mixed
      - insufficient_data
      title: MixedStatus
    NewsProcessingStatusResponse:
      properties:
        news_id:
          type: string
          minLength: 1
          title: News Id
        raw_news:
          anyOf:
          - $ref: '#/components/schemas/ArticleEnrichmentRequest'
          - type: 'null'
        latest_job:
          anyOf:
          - $ref: '#/components/schemas/EnrichmentJobRecord'
          - type: 'null'
        enrichment:
          anyOf:
          - $ref: '#/components/schemas/EnrichmentStoragePayload'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - news_id
      title: NewsProcessingStatusResponse
    NewsResultResponse:
      properties:
        news_id:
          type: string
          minLength: 1
          title: News Id
        raw_news:
          anyOf:
          - $ref: '#/components/schemas/ArticleEnrichmentRequest'
          - type: 'null'
        latest_job:
          anyOf:
          - $ref: '#/components/schemas/EnrichmentJobRecord'
          - type: 'null'
        result:
          anyOf:
          - $ref: '#/components/schemas/ArticleEnrichmentResponse'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - news_id
      title: NewsResultResponse
...
    PipelineStageResult:
      properties:
        stage:
          $ref: '#/components/schemas/PipelineStageName'
          description: Pipeline stage identifier.
        status:
          $ref: '#/components/schemas/PipelineStageStatus'
          description: Final state for this stage.
        fatal:
          type: boolean
          title: Fatal
          description: Whether failure at this stage should stop the pipeline.
          default: false
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
          description: Useful debug message for success, skip, or failure.
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
          description: When execution of the stage started.
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
          description: When execution of the stage finished.
      additionalProperties: false
      type: object
      required:
      - stage
      - status
      title: PipelineStageResult
...
    RawNewsIngestionRequest:
      properties:
        news_id:
          type: string
          minLength: 1
          title: News Id
          description: Unique news identifier.
        title:
          type: string
          minLength: 1
          title: Title
          description: Original article title.
        link:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Link
          description: Canonical article URL.
        ticker:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Ticker
          description: Optional list of related ticker symbols.
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
          description: Publisher or source name.
        published_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published At
          description: Original publication timestamp.
        article_text:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Article Text
          description: Licensed full article text supplied directly by the upstream
            provider.
        summary_text:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Summary Text
          description: Licensed summary/snippet text supplied directly by the upstream
            provider.
      additionalProperties: false
      type: object
      required:
      - news_id
      - title
      - link
      title: RawNewsIngestionRequest
      description: Inbound raw-news payload sent by the upstream backend/news ingestion
        system.
...
    SentimentResult:
      properties:
        label:
          $ref: '#/components/schemas/SentimentLabel'
          description: Discrete sentiment label.
        score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Score
          description: Signed sentiment score between -1 and 1.
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          description: Model confidence between 0 and 1.
      additionalProperties: false
      type: object
      required:
      - label
      - score
      - confidence
      title: SentimentResult
...
    StoragePayloadError:
      properties:
        stage:
          $ref: '#/components/schemas/PipelineStageName'
          description: Pipeline stage related to the issue.
        message:
          type: string
          minLength: 1
          title: Message
          description: Stored failure or warning message.
        fatal:
          type: boolean
          title: Fatal
          description: Whether the issue should be treated as fatal.
      additionalProperties: false
      type: object
      required:
      - stage
      - message
      - fatal
      title: StoragePayloadError
    StoredSentimentPayload:
      properties:
        label:
          type: string
          title: Label
          description: Final sentiment label stored for the article.
        score:
          type: number
          maximum: 100.0
          minimum: -100.0
          title: Score
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
        probabilities:
          additionalProperties:
            type: number
          type: object
          title: Probabilities
          description: Class probability map for the stored sentiment result.
      additionalProperties: false
      type: object
      required:
      - label
      - score
      - confidence
      title: StoredSentimentPayload
    SummaryLine:
      properties:
        line_number:
          type: integer
          enum:
          - 1
          - 2
          - 3
          title: Line Number
          description: 1-based line index.
        text:
          type: string
          minLength: 1
          title: Text
          description: Single summary line.
      additionalProperties: false
      type: object
      required:
      - line_number
      - text
      title: SummaryLine
...
    TickerMixedDetectionResult:
      properties:
        ticker:
          type: string
          minLength: 1
          title: Ticker
        status:
          $ref: '#/components/schemas/MixedStatus'
          description: Overall ticker-level mixed status.
        is_mixed:
          type: boolean
          title: Is Mixed
          description: Whether the ticker sentiment is mixed.
        article_count:
          type: integer
          minimum: 0.0
          title: Article Count
          description: Recent article count used in evaluation.
        lookback_start:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Lookback Start
          description: Start of the recent analysis window.
        lookback_end:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Lookback End
          description: End of the recent analysis window.
        mean_score:
          anyOf:
          - type: number
            maximum: 100.0
            minimum: -100.0
          - type: 'null'
          title: Mean Score
        score_stddev:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Score Stddev
        sentiment_distribution:
          $ref: '#/components/schemas/TickerSentimentDistribution'
          description: Counts of recent article sentiment labels.
        positive_ratio:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Positive Ratio
        negative_ratio:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Negative Ratio
        triggered_reason_codes:
          items:
            $ref: '#/components/schemas/TickerMixedReasonCode'
          type: array
          title: Triggered Reason Codes
        reasons:
          items:
            $ref: '#/components/schemas/MixedReason'
          type: array
          title: Reasons
        thresholds:
          $ref: '#/components/schemas/TickerMixedConfig'
          description: Thresholds used for evaluation.
        recent_articles:
          items:
            $ref: '#/components/schemas/TickerSentimentObservation'
          type: array
          title: Recent Articles
          description: Articles included in the ticker-level evaluation window.
      additionalProperties: false
      type: object
      required:
      - ticker
      - status
      - is_mixed
      - article_count
      - sentiment_distribution
      - thresholds
      title: TickerMixedDetectionResult
    TickerMixedReasonCode:
      type: string
      enum:
      - high_score_variance
      - conflicting_distribution
      - sparse_data
      title: TickerMixedReasonCode
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    XAIPayload:
      properties:
        explanation:
          type: string
          minLength: 1
          title: Explanation
          description: High-level explanation of the enrichment result.
        highlights:
          items:
            $ref: '#/components/schemas/XAIHighlightItem'
          type: array
          title: Highlights
          description: Evidence items grounded in the article text.
      additionalProperties: false
      type: object
      required:
      - explanation
      title: XAIPayload
    XAIResult:
      properties:
        target_label:
          $ref: '#/components/schemas/FinBERTSentimentLabel'
          description: Sentiment label being explained.
        explanation_method:
          type: string
          title: Explanation Method
          description: Backend explanation method identifier.
          default: attention_sentence
        explained_unit:
          $ref: '#/components/schemas/XAIUnit'
          description: Granularity used for highlights.
          default: sentence
        highlights:
          items:
            $ref: '#/components/schemas/XAIHighlight'
          type: array
          title: Highlights
          description: Ordered sentence-level explanation highlights.
        limitations:
          items:
            type: string
          type: array
          title: Limitations
          description: Known interpretation limits for the explanation output.
        sentence_count:
          type: integer
          minimum: 0.0
          title: Sentence Count
          description: Number of sentences included in the explanation scope.
          default: 0
        truncated:
          type: boolean
          title: Truncated
          description: Whether the article was truncated or subsetted for explanation.
          default: false
      additionalProperties: false
      type: object
      required:
      - target_label
      title: XAIResult
    XAIUnit:
      type: string
      enum:
      - sentence
      title: XAIUnit